Calculated the f1_score between output of padding and blender
Made a mistake in the previous blender. Calculated the f1_score between the output of union of CNN and LSTM and the blender, instead of between output of padding and blender. This seems to be the best basic for comparing the f1_score with those of CNN and LSTM. 

This blending ensemble improved LSTM's f1_score by 0.03